+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * NEW: Add ability to swap in a custom transport. Adds `Raven.setTransport`, and `transport` option to config. Docs: https://docs.getsentry.com/hosted/clients/javascript/config/
 * CHANGE: Always expose `Raven` to `window`. Please call `Raven.noConflict()` if you want it restored to what it was. See: https://github.com/getsentry/raven-js/pull/393
 * DEPRECATED: Replace `Raven.setReleaseContext` with `Raven.setRelease`.
+* NEW: Add `Raven.clearContext()` to empty all of the context.
+* NEW: Add `Raven.getContext()` to get a copy of the current context.
 
 ## 1.1.22